requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.0
black==20.8b1
click==7.1.2
Django==3.1.3
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
regex==2020.10.28
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
